% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_hex.R
\name{cor_hex}
\alias{cor_hex}
\title{Correlation Hex Plot Function}
\usage{
cor_hex(
  dt = dtplot,
  id.col = "protein",
  cor.method = c("pearson", "kendall", "spearman")[1],
  savefile = "outputfile",
  singleplotsize = c(3, 2.5),
  facetplotsize = c(3 * 3, 2.5 * 3),
  bin = 50
)
}
\arguments{
\item{dt}{A data frame containing the data to be analyzed.}

\item{id.col}{The name of the ID column, default is "protein".}

\item{cor.method}{The correlation method to use, default is "pearson".}

\item{savefile}{The directory to save plots, default is "outputfile".}

\item{singleplotsize}{Size of each individual plot as a numeric vector of width and height, default is c(3, 2.5).}

\item{facetplotsize}{Size of facet plot as a numeric vector of width and height, default is c(9, 7.5).}

\item{bin}{Number of bins for hex plot, default is 50.}
}
\value{
A list containing facet plot and individual plots.
}
\description{
This function generates correlation hexagon plots for a given data frame.
}
