% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{Example Dataset for corhex}
\format{
A tibble with 847 rows and 7 variables:
\describe{
\item{Index}{Character. Unique identifier for each protein, such as "protein_847".}
\item{replicate1}{Numeric. Expression levels of the protein in the first replicate.}
\item{replicate2}{Numeric. Expression levels of the protein in the second replicate.}
\item{replicate3}{Numeric. Expression levels of the protein in the third replicate.}
\item{replicate4}{Numeric. Expression levels of the protein in the fourth replicate.}
\item{replicate5}{Numeric. Expression levels of the protein in the fifth replicate.}
\item{replicate6}{Numeric. Expression levels of the protein in the sixth replicate.}
}
}
\usage{
data(demo)
}
\description{
This dataset is a small example of protein expression data
designed to demonstrate the functionality of the \code{cor_hex} function.
}
\examples{
data(demo)
head(demo)

}
\keyword{datasets}
