% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corhex.R
\name{cor_hex}
\alias{cor_hex}
\title{Correlation Hex Plot Function}
\usage{
cor_hex(
  dt = dtplot,
  id.col = "protein",
  cor.method = c("pearson", "kendall", "spearman")[1],
  savefile = "outputfile",
  singleplotsize = c(3, 2.5),
  facetplotsize = c(3 * 3, 2.5 * 3),
  bin = 50
)
}
\arguments{
\item{dt}{A data frame containing the data to be analyzed.}

\item{id.col}{The name of the ID column in the data frame. Default is "protein".}

\item{cor.method}{The correlation method to use. Options are "pearson" (default), "kendall", or "spearman".}

\item{savefile}{Directory to save the generated plots. Default is "outputfile".}

\item{singleplotsize}{Size of each individual plot in inches (width, height). Default is c(3, 2.5).}

\item{facetplotsize}{Size of facet plot in inches (width, height). Default is c(9, 7.5).}

\item{bin}{Number of bins for hex plot. Default is 50.}
}
\value{
A list containing two elements:
\itemize{
\item facet_plt: Facet-wrapped correlation hexagon plot.
\item singleplot: List of individual correlation hexagon plots.
}
}
\description{
The function calculates pairwise correlations between numeric columns in the input data frame \code{dt},
and generates hexagon plots showing these correlations. It also includes facet-wrapped plots for each
unique pair of variables, with correlation coefficients annotated on the plots.
}
\details{
This function generates correlation hexagon plots for a given data frame.
}
\examples{
dt <- data.frame(
  protein = c("A", "B", "C"),
  var1 = c(1, 2, 3),
  var2 = c(4, 5, 6),
  var3 = c(7, 8, 9)
)
cor_hex(dt)

}
